refactor: 게시글 상세 컴포넌트 분리
